Create new file for abstract models
